sender/pyproject.toml
[tool.poetry]
name = "sender"
version = "0.1.0"
description = ""
authors = ["Piotr Szczygieł <szczygiel.p@gmail.com>"]

[tool.poetry.scripts]
sender = "sender:run"

[tool.poetry.dependencies]
python = "^3.7"
kivy = "1.11.1"
"kivy_deps.glew" = "0.1.12"
"kivy_deps.sdl2" = "0.1.22"
pillow = "^6.2.1"
segno = "^0.3.2"
numpy = "^1.17.4"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
mypy = "^0.750"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
